Change caching (#1)
Change caching because the original way prevented requests from being
tracked. With Nginx caching, requests would never get back to the
application so that it could log the request.  The new way uses
cacheutils and a small in-memory TTL cache simply to prevent floods of
repeat requests from causing a performance problem.