Fix errors for fields with array values
Fix server error given a `fields` request parameter for a field that has
arrays of objects as its values. Say
`field=sourceResource.language.name`, where we might get [{'name':
'English'}] depending on how it was ingested and what the provider gave
us.

Before this fix, we would have expected the 'language' property to be an
object with a key called 'name', but we've since found that it can be an
array of objects with the key 'name'.